feat(a4): establish workout component flow
